103 refactor after code review and fix (#104)
* 로그아웃과 회원탈퇴 방어적 코드 삽입

* HTTP 코드 관련 수정

* 서비스 로직 순서 검증 관련 수정

* 경쟁 조건 관련 수정

* n+1 수정(일괄조회,인메모리 멥핑)

* 수정

+++ src/main/java/com/shelfeed/backend/domain/auth/controller/AuthController.java
@@ -3,6 +3,8 @@ package com.shelfeed.backend.domain.auth.controller;
 import com.shelfeed.backend.domain.auth.dto.request.*;
 import com.shelfeed.backend.domain.auth.dto.response.*;
 import com.shelfeed.backend.domain.auth.service.AuthService;
+import com.shelfeed.backend.global.common.exception.BusinessException;
+import com.shelfeed.backend.global.common.exception.ErrorCode;
 import com.shelfeed.backend.global.common.response.ApiResponse;
 import com.shelfeed.backend.global.jwt.JwtProvider;
 import com.shelfeed.backend.global.security.CustomUserDetails;
@@ -13,6 +15,7 @@ import jakarta.validation.Valid;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseCookie;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
@@ -27,12 +30,14 @@ public class AuthController {
 
     // 1. 이메일 회원가입  POST /api/v1/auth/signup
     @PostMapping("/signup")
-    @ResponseStatus(HttpStatus.CREATED)//성공 시 상태
-    public ApiResponse<SignupResponse> signup(
+    public ResponseEntity<ApiResponse<SignupResponse>> signup(
             @Valid @RequestBody SignupRequest request,HttpServletResponse response) {// 유효성 검사 + JSON 형태의 요청 본문 DTO로 변환
                 AuthService.TokenPair result = authService.signup(request);
                 setRefreshTokenCookie(response, result.refreshToken());
-                    return ApiResponse.success(201,"회원가입이 완료되었습니다. 이메일 인증을 진행해주세요.", result.response());
+                HttpStatus status = HttpStatus.CREATED;
+                return ResponseEntity
+                        .status(status)
+                        .body(ApiResponse.success(status.value(),"회원가입 완료", result.response()));
     }
 
     // 2. 이메일 인증코드 확인  POST /api/v1/auth/email/verify
@@ -93,6 +98,9 @@ public class AuthController {
             @RequestHeader("Authorization") String bearerToken,
             @CookieValue(name = "refreshToken", required = false) String refreshToken,
             HttpServletResponse response) {
+        if (bearerToken == null || !bearerToken.startsWith("Bearer ")){
+            throw new BusinessException(ErrorCode.INVALID_TOKEN);
+        }
         String accessToken = bearerToken.substring(7);
         authService.logout(userDetails.getMember().getMemberUserId(), accessToken, refreshToken);
         deleteRefreshTokenCookie(response);
