+++ src/main/java/com/shelfeed/backend/domain/member/controller/MemberController.java
@@ -11,6 +11,8 @@ import com.shelfeed.backend.domain.member.dto.response.UpdateGenresResponse;
 import com.shelfeed.backend.domain.member.dto.response.UpdateProfileResponse;
 import com.shelfeed.backend.domain.member.dto.response.UserProfileResponse;
 import com.shelfeed.backend.domain.member.service.MemberService;
+import com.shelfeed.backend.global.common.exception.BusinessException;
+import com.shelfeed.backend.global.common.exception.ErrorCode;
 import com.shelfeed.backend.global.common.response.ApiResponse;
 import com.shelfeed.backend.global.jwt.JwtProvider;
 import com.shelfeed.backend.global.security.CustomUserDetails;
@@ -18,6 +20,7 @@ import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseCookie;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
@@ -33,10 +36,11 @@ public class MemberController {
     private final JwtProvider jwtProvider;
     // 1. 온보딩  POST /api/v1/users/me/onboarding
     @PostMapping("/me/onboarding")
+    @ResponseStatus(HttpStatus.CREATED)
     public ApiResponse<OnboardingResponse> completeOnboarding(
             @AuthenticationPrincipal CustomUserDetails userDetails,
             @Valid @RequestBody OnboardingRequest request) {
-        return ApiResponse.success(200, "온보딩이 완료되었습니다.",
+        return ApiResponse.success(201, "온보딩이 완료되었습니다.",
                 memberService.completeOnboarding(userDetails.getMember().getMemberUserId(), request));
     }
 
@@ -97,8 +101,11 @@ public class MemberController {
     @DeleteMapping("/me")
     public ApiResponse<Void> withdraw(
             @AuthenticationPrincipal CustomUserDetails userDetails,
-            @RequestHeader("Authorization") String bearerToken,//요청헤더에서 Authorization 만 가져오기
+            @RequestHeader(value = "Authorization", required = false) String bearerToken,//요청헤더에서 Authorization 만 가져오기
             @RequestBody WithdrawRequest request, HttpServletResponse response){
+        if (bearerToken == null || !bearerToken.startsWith("Bearer ")){
+            throw new BusinessException(ErrorCode.INVALID_TOKEN);
+        }
         String accessToken = bearerToken.substring(7);
         memberService.withdraw(userDetails.getMember().getMemberUserId(), accessToken, request);
         deleteRefreshTokenCookie(response);
